backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
